Added schema and index files for ICDC.
